added an original GUI
idk why but after doing ROI warping all the functions that involves affine transform became weird in the middle of the transformation, though the results seem OK, so might still need debugging or I have to think of another approach to somehow freeze the body and the bg as most as I could. The body might be easy to freeze cuz I have the segmentation mask, but the bg is almost impossible, I should probably try to freeze the body first and just tell them this is the most I could do. gl to me