Now the warping should be fairly limited
decreasing number of landmarks on the eyes does not work, but I observed that only cases with small eyes tend to have this issue. Also the function for enlarge mouth has this issue as well. So maybe try upscaling the face? But need one more detector.

I found this useful resource: http://www.gson.org/thesis/warping-thesis.pdf
I will try to implement those functions in this way, a radical approach in mesh warping, if it works, I could even change the whole piecewise affine to this one.